+++ src/L2Data2Code/L2Data2Code.BaseGenerator/Configuration/IGlobalsConfiguration.cs
@@ -1,5 +1,7 @@
+using L2Data2Code.BaseGenerator.Entities;
 using L2Data2Code.SharedLib.Configuration;
 using Newtonsoft.Json.Linq;
+using System.Collections.Generic;
 using System.Collections.Specialized;
 
 namespace L2Data2Code.BaseGenerator.Configuration
@@ -8,5 +10,6 @@ namespace L2Data2Code.BaseGenerator.Configuration
     {
         NameValueCollection Vars { get; }
         NameValueCollection FinalVars { get; }
+        List<FinalCondition> FinalConditions { get; set; }
     }
 }